2024-11-01 PM

+++ braincare/src/main/java/org/zerock/braincare/todos/domain/Todo.java
@@ -18,7 +18,7 @@ public class Todo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int todoId;
+    private Long todoId;
 
     private String title;
